fixed: not skipping processing erroneous script
That is, a specific error, when identified in a node script,
was not causing the program to stop processing the script
as it should (even though the error was being properly
caught and reported).

This error was actually caused by a "continue" statement that
was not indented correctly (actually it should have been dedented
a bit).