added: new snippet nodes (useful encapsulations)
Added 04 new snippet nodes. The first 03 emulate the "obj[item]"
syntax just like the "a[b]" operator node. They are the "get_at_int",
"get_at_string" and "get_at_literal". The only difference from the
operator node is that they provide specific widgets to help you
specific the item to be picked.

The last node added was the "namespace_from_exec", a node that allows
you to provide a custom script to be executed and returns the resulting
namespace as a dict.