fix: deleting subparam widget crashes app
That is, in the following case: when a subparam widget is deleted and
the node has a regular parameter with a widget the app was crashing.
This was so because we were mistakenly treating a widget instance as
a dict.

This happened because I wasn't making sure the parameters searched
were of variable-kind. Now that I fixed this the process resumes
without problems.