move: 📁 "aww" folder (contains in-app hypertext)
📂nodezator
│
└── 📂data
    │
    ├── 📂locale
    │   └── 📂en_us
    │       └── 📁aww    <-- this folder
    │
    ├── (📁aww)          <-- is now here
    │
    └── [...]

The folder called "aww", containing hypertext used for in-app content
like the manual, controls page, etc., was moved to the top of the "data"
folder, now sitting at "data/aww" location. This "aww" folder contains
folders that each work like a local website.

Before, the "aww" folder was inside the "data/locale/en_us/" folder,
specific to English content. Not desirable, because we would need a
different "aww" folder for each language the "locale" folder
(for instance, we would need a "locale/pt_br/aww" folder). For text
content this would be ok, but since the local websites inside the "aww"
folder also contain images that can be used for many different
languages, we didn't want to duplicate those images as well when
translating the local websites into other languages.

Because of this, we moved the "aww" folder outside the "locale" folder
and now the local websites inside it can handle language-specific
content by themselves, just like a normal website on the internet and
reference the same images accross different languages.

For instance, for the Nodezator's manual, we would need to do this
before:

📂locale     (note there's more than one "aww" and
│             "manual.nodezator.website" below, one for each
│             language)
│
├── 📂en_us
│   │
│   └── 📂aww
│       │
│       └── 📂manual.nodezator.pysite
│           │
│           ├── 🗎 index.htsl
│           └── 📂images
│               ├── 🗎 img00.png
│               └── 🗎 img01.png
│
└── 📂pt_br
    │
    └── 📂aww
        │
        └── 📂manual.nodezator.pysite
            │
            ├── 🗎 index.htsl        (text content would be
            │                        translated - no duplication,
            │                        that's ok)
            │
            └── 📂images          (images, or at least some
                │                  of them, would be duplicated,
                │                  not ok; here "img00.png" and
                ├── 🗎 img00.png    "img01.png" would both be
                └── 🗎 img01.png    duplicates of the images at
                                   "en_us/aww/manual.nodezator.pysite")

Now language-specific content for each website is handled inside the
websites themselves, just like in websites on the web:

📂aww        (now we have a single "aww" folder, and a single folder
│             for each website; all content for a single local
│             website, including different translations, are now
│             managed inside the website itself)
│
└── 📂manual.nodezator.pysite
    │
    ├── 🗎 index.htsl
    ├── 📂images
    │   ├── 🗎 img00.png
    │   └── 🗎 img01.png
    │
    └── 📂pt_br           (each additional language can have
        │                  its own folder now; the website structure
        │                  is up to to us, actually; the important
        │                  thing is that we don't need to duplicate
        │                  content, since it is all available at the
        │                  same location)
        │
        └── 🗎 index.htsl  (for instance, pt_br/index.htsl can
                           now reference images from "../images", like
                           "img00.png", since they are located on the
                           same website, no need to duplicate)