change: easy to provide new paths on file browser
It is now easier to provide new paths (files and folders) on the
file browser. Before, in order to type the name for new paths
(usually files) in the file browser, it had to be invoked in its
"create_path" mode. This mode made an entry available at the bottom
of the file browser for the user to type the name of a new path.

This is fine, but the "select_path" mode (another one available) was
far more used and it didn't provide such entry, making it so the user
had to click specific buttons in the file browser to invoke a form
to enter the name for a new path and hit the button "return new path".
The way things were handled in "select_path" mode were convoluted and
unintuitive.

Now both modes were merged together, making things easier and more
intuitive. The entry at the bottom of the file browser is now always
available and can be used to provide name(s) of new path(s) as desired.
The user will probably use it mostly or exclusively to provide the name
for a single new file or folder, though. For instance, when saving
a new file or exporting a file into a new one with another format.

### this is the commit message draft
###
### In order to not forget what I did, I write the commit message
### here in advance (commit subject and description) as I work.
###
### Then, when it is time to commit, I submit this file as the
### commit messsage using one of the options below:
###
### -t : to edit before committing
### -F : to use as-is