change: use empty temp file when creating new one
That is, instead of prompting user for a path when creating
a new file, just create and load a temporary untitled empty file
automatically.

This way the user can start editing new files right away without
worrying about where to save the file and even whether they'll end
up doing so (maybe they just created the new file to
experiment).

Only when the file is saved the first time the user is prompted
to provide a path.

As insightfully suggested by "Tom999#7817" on discord (same as
@Tom-99 on github), mentioned in discussion #33.