change: import pygame services from custom spot
That is, instead of being imported from pygame directly,
some key pygame services like input-related functions
and screen updating are now imported from a custom spot,
which in this case is the pygameconstants.py module.

This means that, if we desire, we can now write custom functionality
that extends the pygame functionality and provide it instead.

This opens up many possibilities, like the implementation of
input recording/replay for automated GUI testing and usage
demonstrations (something like a demonstration/tour mode).