fix: clear event queue on loop holder switching
That is, we clear the event queue whenener we are about to switch
to a new loop holder to prevent events generated in the current
loop holders to spill into the new one.

Even though we present a reasoning for this change, events from
a loop holder should never spill into the next one, at least in
theory. More precisely, since we do not inject events and actually
grab all events directly from pygame.event.get(), we suspect this
behavior to be a bug in the library itself.

To reproduce the bug, just checkout the previous commit and,
after launching nodezator, leave the splashscreen and press
<Ctrl-N>. On top of the KEYDOWN event generated, which will
create a temporary file automatically, 11 new equivalent KEYDOWN
events will be generated, which will try to repeat the operation
11 times, causing unexpected behaviors like a dialog popping up
many time.

The version used was pygame 2.1.2 with SDL 2.0.16 and Python 3.10.6