fix: remove erroneous unneeded checks/calls
There was no need to check for the existence of both the
temp swap path and the swap path referenced in APP_REFS.swap_path.

Just checking the existence of the APP_REFS.swap_path attribute
is enough, since it reference the swap path used for both regular
files and temporary (untitled) files. Whichever is in use will
be properly removed, which is the intended outcome.