misc: key pygame service imports use namespace
Imports of key pygame services that control the app mode, like
callables to get events, key pressed states, etc. were changed
so a namespace is imported instead of the callables themselves.

The callables are now accessed as attributes in such namespace.

A new mainloop exception was introduced that causes the app to
reset its state.

Such changes are intermediary, made in other to allow the
different modes (normal, input recording and input playing) to
be activated at will while the app is running, instead of having
to pick a single mode on launch.

To finish the changes, we still need to change the pygamesetup
subpackage to create the namespace and allow it to switch behaviours
depending on the activation of different modes.