remove: let pygame set DEPTH automatically
That is, rather than manually setting the value of depth for
pygame.display.set_mode, we ommit it completely, which makes it
so pygame sets it for us.

According to the online documentation for the pygame.display module:

"It is usually best to not pass the depth argument. It will default
to the best and fastest color depth for the system."

https://www.pygame.org/docs/ref/display.html