change: reorganize pygamesetup services
The services were reorganized so that it is easier to switch
between different modes (normal, recording and playing), though
the actual switching wasn't implement yet.

That is, the changes are still unfinished, but the app is already
launched again, in normal mode only. The next changes should bring
back the input recording and playing modes, this time more organized
and refactored.