remove: command line options
That is, removed the most recent command line options, the ones
related to the new input recording and playing modes.

This is so because such modes will be accessible in the
graphical interface soon. In other words, the next changes are
meant to make it so the app can switch between different modes
without having to set them on startup.