change: use TEXTINPUT events for typing text
Rather than KEYDOWN events, we now use TEXTINPUT events instead for
capturing text typed on the keyboard.

One of the advantages is reducing the quantity of events that we
need to keep track, because TEXTINPUT (and the related TEXTEDITING)
events can be turned on and off using the start_text_input() and
stop_text_input() functions from the pygame.key module, respectively.

Another advantage is that this makes it easier to work with IME
editors/input methods for other languages.

We also set the rect used for typing with the IME editor (in case one
is being used), but ensuring it is accurate and that is properly
updated when the window is resized must be done in another commit.

This will most likely be postponed, so we can test this properly in
multiple environments.