change: compress recorded session data
Did so by removing unrelevant event data from the recording and
by notifying users before starting recording whenever the caps lock
and num lock keys (toggle keys) are turned on. It is okay to leave
them on, but if they are not necessary, then it is better to turn them
off because it results in less data to be recorded.

Also record name of TEXTINPUT event as 'ti', which is more compact.