change: compress recorded session data
This was done by shortening how regular and modifier key presses
are stored. We actually just reversed the maps: instead of storing
frame numbers as keys pointing to the pressed regular/modifier keys,
the pressed regular/modifier keys are used as keys and point to the
frames wherein they are pressed.

As always there's no loss of data since that data is changed back
into its full format whenever we need to play the session data.