change: close on quit when not in normal mode
That is, when not in normal mode, whenever the user attempts to
close the app by clicking the "x" icon on top of the window, close
the app immediately.

In both record and play mode, the focus is on recording or playing,
not in file editing. That is, the app is not being used normally.
Thus, when the user attempts to close the app, it should close
immediately, regardless of whether there are unsaved changes or not.

In other words, since the unsaved changes outside normal mode are
the result of staged/planned actions (when in recording mode) or
pre-recorded actions (when in playing mode), they are less meaningful,
more volatile, in a way.

They represent real work, real actions, real events, yet I do not
consider them to be real work per se. Simply put, changes made outside
normal mode are means to record or play a session rather than being
the purpose themselves, like they are in normal mode.

In normal mode, I make and want to keep the changes, so quitting the
app considers the existence of changes and adds additional measures
to prevent loss of those changes by asking user to save if there are
unsaved changes, etc. In the other modes the changes are not important
and thus are promptly ignored.