add: display session duration/indicate progress
That is, when in playing mode, we display the approximate session
duration (using a label) and also indicate the current progress
by using a thin horizontal bar similar to the one used in youtube
videos. But ours is displayed at the top of the application window,
not at the bottom.