change: speed label text and speed precision
That is, changed the label on the "max" speed button on the form to
set a playing session. It now reads "max/uncapped", which explains
a bit more.

Also changed the integer/float entry used to specify the speed to
accept floats, so the user can specify even slower framerates, the
ones between 0 and 1. I'm not sure if this will be needed, but it
is a change that has no cost or negative impact.

Even if a user sets a speed too slow, all that has to be done is to
cancel playing the session, then go back to the form and set an
appropriate speed. In more than 99% of the cases, though, the user
will probably be satisfied in using one of the speed rates shown in
the speed buttons. Most people won't probably even notice that a
speed rate between 0 and 1 can be used. And even so, there will be
no harm in it.