change: use our subpackages for specific services
The services in question are the definition of objects for general
use and text surface rendering. The changes in this commit concern
the usage of such services for creating and storing labels for the
recording and playing modes.

At first, we used anonymous objects and local system fonts for the
labels, but we already have our own custom text rendering facilities
and our favorite font as well, so I believe it is better to use them.