change: improve manual and license info
The manual was improved to better guide the users on the process
of loading and distributing node packs, with several changes in the
corresponding chapters. The changes were made in response to the
feedback provided by @arthexis (Mr. Guillen) on github discussion #45.

The manual was also updated to reflect the recent fix in the mechanism
to present and solve problems in local node packs when loading files.
That is, the fix represented by the previous commit.

The license info was also improved with better wording only, no changes
in the licensing were made.