misc: improvements to python exporting feature
The "additional levels" optional setting in the form was removed,
since its advantages are very limited and its usage was assumed to
be rare (and removing it won't break anything).

As such, there's no need for a form anymore and the user can
export the graph as a Python file right away since the file manager
is now presented instead of a form.

A new command was added to the "File" menu, entitled "View as Python".
It allows the user to view the exported code directly in the app
without the need to save it. This is very convenient since before,
the user had to save the exported code in a file and open it with an
external app to be able to see the exported code.

The new command can be accessed with the keyboard with Shift+Ctrl+P.