fix: include toggle button in image exporting
Forgot to take signature mode toggle button into account when
exporting the layout as an image. This button is the one that
appears in the topleft corner of callable nodes (user-defined nodes,
built-in nodes, standard library nodes and snippet nodes) when
not in callable mode. The one that allows the user to quickly
toggle between expanded and collapsed signature mode.