remove: unnecessary if-block
Removed a needless if-block used to check whether a callable node had
outputs before adding the output rects manager to the general
rects manager used in the expanded mode.

Although all callables have outputs, I added this check preemptively
in anticipation of a future feature that would hide the output sockets
from the node (the collapsed mode does that to nodes that support it,
but it is unrelated). This feature I anticipated won't hide sockets
anymore, so this if-block ended up being needless.

In summary, I removed the if-block and included the rects manager
directly in the list of rects managed by the more general rects
manager.