add: 02 new snippet nodes
The nodes are return_untouched and for_item_in_obj_pass.

The return_untouched node is an identity function.

The for_item_in_obj_pass snippet is used to perform an "empty" for-loop
(containing just a "pass" statement) with the items of an iterator.
Its purpose is to trigger the evaluation of the items from an iterator.

For instance, a "map(print, range(5))" object won't print the values
in the range by itself, since the items are calls are performed lazily.
To solve this, just pass such map object to the "for_item_in_obj_pass"
node and all the calls will be performed one by one.

Of course you could also pass that map object to a "list" node or
other collection, but it would create and populate a collection
with the output of the calls. But, this would be desirable only
if you were interested in the outputs. In the case of our map object,
however, since it calls the print function, which always returns None,
you would end up with a collection full of None objects.

The "for_item_in_obj_pass" node on the other hand ignores the items.
Its only purpose is to trigger the execution.

In other words, it is very useful when you want the items of an
iterator to be evaluated but don't care about the outputs. That is,
when you are only interested in the side-effects.