change: move output sockets on top of nodes
That is, output sockets are now displayed on top of the following
nodes:

- user-defined nodes;
- built-in nodes;
- standard library nodes; and
- snippet nodes.

This is actually the usual way to position outputs in nodes in other
apps as well.

I also plan to use the area to the left of nodes, just below the
last output socket, to display in-graph visualization in the new
viewer nodes to be implemented next.