add: method to ensure arguments are cleared
That is, after executing the graph.

Before this, the app was already doing a pretty good job of clearing
such arguments but there was a case which was missing this treatment.

A new execution method that executes only upstream nodes from a
viewer node and the viewer node itself, in order to update its
visualization, also benefits from this new method.