add: checkered background on preview panel visual
That is, we ensure the surface used as the visual on preview panels
is first blit over a checkered surface created on the spot (it is a
copy of a cached instance).

Such newly created surface is then used instead of the original
visual. This will only have an effect when the original surface
has transparency, in which case the checkered pattern will show.

This is desired, since it prevents the surface set on the
preview panel to have real transparency, which would harm the
app's performance. Morever, a checkered pattern is a nice and
established way to convey transparency anyway.