fix: intfloatentry leaving mouse invisible
When the user clicks the intfloat entry, it automatically enters
a standby mode which waits for one of two events to define which
mode it must enter next.

If the event is a MOUSEMOTION, that is, the movement of the mouse,
it enters mouse edition mode (changing the value by dragging the
mouse). If it is a MOUSEBUTTONUP for the mouse left button, that is,
the mouse left button being released, it enters keyboard edition mode
(changing the value by deleting/typing characters in the keyboard).

The problem was that after setting up the new mode, instead of
ignoring the next events in the queue, we were still processing them.

On top of being unnecessary, this was also harmful, because it meant
that we were risking setting up the same mode more than once or
setting up another mode.

Add to that the fact that in this context, when a new mode is set,
we do not execute proper exit setups for the previous one, cause we
assume the previous one was just the standby mode being left behind,
and which needs no exit setups.

This was particularly nasty if we entered the mouse edition mode
and then enter the keyboard edition mode immediately after. This is
so because the mouse edition mode makes the mouse invisible (so it
doesn't distract the user while dragging the mouse to change the
value in the entry) and the mouse edition mode isn't properly exited
when switching to keyboard edition mode, so the mouse visibility is
never turned back on, causing it to remain invisible afterwards.

This was fixed, and now when we switch from the standby mode to the
next one, regardless of whether it is the mouse edition or keyboard
edition mode, we ignore the current events in the queue by exitting
the for-loop with a "break" statement, preventing a new mode to be
accidentally set.