add: surface viewer
New viewer to visualize surfaces on screen. Uses old code from when
the images previewer in imagesman/previewer was imagesman/viewer and
had a full mode in which a surface representing an image in full size
could be displayed on the screen.

This new one is more general and can display any surface, not only
those representing images on disk.