fix: grab coordinates from screen rect directly
In widget/intfloatentry/modes.py we were storing coordinates from the
screen rect at the beginning of the module as constants for reuse.
This is wrong because the size of the screen may change during a
session, in which case the coordinates will be different.

Now, instead, we grab the values from the screen rect directly, which
is updated whenever the window changes size.

The original behaviour was fine when Nodezator's window was of a fixed
size in the past, but from the moment we allowed the user to change
the size we should have updated this accordingly.