fix: prevent initial big jump in intfloat value
In intfloat entries, after entering mouse edition mode (when the
mouse is dragged around to change the value) we make the mouse
invisible and keep it near the center of the screen, since it is
easier to keep track of its movements when it is there, specially
if the user makes long dragging movements.

We measure such movements using mouse motion events. However, the
first one or two mouse motion events might have been generated before
setting the position of the mouse to the center of screen. When this
happens, the mouse position stored in the events might be too distant
from its current position in the center of the screen, causing a big
jump in the value which was not caused by the user's movement and
therefore is undesirable.

Because of that, we introduced this fix, which watches out for such
big jumps and ignores them. It is also efficient, since it only does
so at the beginning of the dragging movement and once a normal jump
in the value is detected, it falls back to a regular routine which
doesn't need to make such extra check.