add: .md and .pyl files in source distribution
A .pyl file is just a Python file containing a Python literal. That is,
its text can be passed to the literal_eval() function from the ast
standard library module.

It is a format used to hold data just like JSON files, but since it uses
Python syntax it can do things like having integers used as dictionary
keys, for instance.

.md (markdown) files are used in the source code for conveying
information about the code but also for doctests.