test: update doctest in 📁graphman/validation
The doctest file for the check_return_annotation_mini_lang() function
in the graphman/validation directory was updated to use the
markdown (.md) format instead of Python (.py).

The test data that was previously stored in a Python file now uses
the .pyl format (which is just a Python file containing a Python
literal whose source can be passed as a string to the literal_eval()
function of the ast standard library module.