add: improve creation of "not found" icon
Drawing the icon in a surface or representing it in svg now
results in a more accurate drawing regardless of the dimensions
of the surface, that is, whether the width is the same as the
height or either of them is longer.

Before, the line segment crossing the ellipse would be a bit off
depending on the width and height used and/or whether one was
longer than the other.

For instance, the icon represented as svg in preview panels (their
size is 256x256) would feature a line segment that went beyond the
ellipse's outline.