fix: NameError in Python exporting edge case
When exporting a graph to Python, if the graph has no nodes,
it will end up raising a NameError due to a missing definition
in the Python exporting function.

More precisely, a function used as a key in a call to sorted() was
missing its definition. The definition was probably accidentally
lost in one of the refactorings made before uploading the repo to
github.

It is rare for the if-block that was using this missing definition
to be executed, because it is only entered when the user tries to
export the graph when there are no nodes in it. You might be asking
why this is even possible. The answer is that there might still be
text blocks in the graph and their text is used as comments in the
exported Python code.

Because of that, the condition in the if-block was also updated so
it only enters when there are text blocks in the graph (unless
there's only one and it is close to the origin of the graph, in
which case it is used as the docstring of the exported function).