change: set proper "not found" surfaces/colors
In other words, we eliminated inconsistencies in the interface by:

- using "not found" surfaces with the right colors where needed;
- using proper "not found" colors for SVG representations.

Before these changes, there were some surfaces on which we were
drawing the "not found" icon with the wrong colors (temporary
"placeholder" colors). Now we generate cached surfaces with the
"not found" icon on them, which already come with the right colors.

The SVG representations were also using "placeholder" colors and
were fixed as well.