change: rename HTSL Browser to Local Pages Browser
That is, the title that appears at its top.

This way its purpose is more clear, since HTSL is an unknown term.

HTSL is just a custom XML format very similar to HTML that we use
in Nodezator to define hypertext that is displayed in Nodezator.

There's no need to use this term, that is, HTSL, where regular
users can see.

I also moved this title to appear after the current page's title.
This way the user sees the page title's first, which is more useful.