change: some colors of local browser
Two elements of the local browser had their colors changed to more
pleasing and neutral colors. Such elements were the title at the top
which is now black (almost black actually) and the background of the
browser (not the portion where the page is displayed, but the outside
borders) which is now a medium grey.