add: horizontal scrolling option in local browser
That is, an alternative way. The pages in the local pages browser can
already scroll horizontally, but we added a new method: by pressing
the shift key while using the mouse wheel.

The local pages already scrolled vertically with the mouse wheel alone,
so it is only natural that we implement the horizontal movement with
the shift key (just like web browsers do).