add: better numbering and spacing for ol/ul tags
That is, better numbering for nested ol tags, meaning items in nested
ordered lists are prefixed with the numbering from a parent ordered
list.

The improvement in spacing affects both ol and ul tags when they are
nested (regardless of whether it is a nesting solely of ol lists,
solely of ul lists or a mix of both) and it means that extra space
is inserted above and below a nested list, so that is isolated from
sibling items and thus becomes easier to spot, specially in large
nested lists like those used to represent a table of contents.