change: line numbering for preformatted text
That is, changed how line numbering for preformatted text is set
in the local pages browser.

Python code blocks (defined with the "python" tag) can now be
displayed without line numbering, which is the new default
behaviour. To enable line numbering, one must use the new "uselineno"
attribute of the tag, set to "true".

General preformatted text (defined with the "pre" tag) already had this
behaviour, but we renamed its "use_lineno" attribute to "uselineno".