change: remove commented out nodes from given ones
That is, whenever executing a given set of nodes (rather than all
of the existing ones), first filter out the commented out ones
from them, executing only the remaining ones, if any.

The ability to execute only a smaller set of the existing nodes is
used when we trigger execution of a viewer node and its upstream nodes
by clicking the viewer node's reload button.