fix: width of some nodes' rects
That is, the width of the rect from the "rect" attribute of nodes
which don't have inputs.

Before, when such a node was instantiated, the rect's width only
covered the space between the node's lateral edges, which wasn't
enough because when such rect was used to draw a selection rect
around the node, the selection rect was having its sides hidden
behind the node.

Now this isn't a problem anymore because this is being taken into
account in the calculations, along with other possible factors.