change: rename htap protocol to nodezator protocol
The htap protocol is not a fancy protocol as it seems. It is just
a way to indicate that a link points to local content in the app rather
than content from the world wide web. It means "hypertext from app" or
something like that.

Thus, to avoid confusion, I'm renaming htap to nodezator in all links
referenced in the package. Instead of seeing "htap://" at the beginning
of links, you'll now see "nodezator://". And I believe it is now a bit
easier to see that it just refers to content that is found in the app
itself.