change: expected type in some widgets
Make it so preview widgets all use string ("str") as the expected
type. Also make it so color button widgets that use hex strings
use string as the expected type.

For now this is done in order to avoid confusion, since string
is the type primarily returned by those widgets, even though they
can also return a tuple with multiple strings when more than one
value is set on them.

This should be revisited in the future when we implement type
checking, an optional feature for the users to enable and configure
as they wish.