add: perform_attr_call snippet node/encapsulation
A new snippet node that represents the act of retrieving an
attribute from an object and performing a call to it with
given arguments.

Created as a convenience so users don't need to use a getattr()
and a perform_call() node every time they want to call an object's
method. Instead, this perform_attr_call() node already encapsulates these
actions.