change: scroll objects only once per frame
That is, scrolling/moving the canvas, which is equivalent to moving
the objects in it is now is only performed once per frame at most.

Before, the scrolling could be performed twice, because it was
performed vertically, then horizontally. This two-step scrolling
isn't needed and wastes time, because we can simply scroll (move
the objects) in both axis at once, which we are doing now after this
change.