fix: error when executing collapsed nodes
The error was caused by using the wrong attribute when retrieving
the existing sockets from a collapsed node. Using the correct
attribute name fixed the error.

In summary the error was causing the output of nodes to be handled
in the wrong way because it wasn't able to accurately determine the
number of existing output sockets. This problem could cause all sorts
of other problems like the graph failing to execute or the wrong data
being passed along to other nodes.