fix: error when executing operation nodes
The error was caused by using the wrong attribute when retrieving
the existing sockets from an operation node (which are always a
single socket). Adding an if/else clause that uses the proper
attribute when it is needed fixed the error.

The error was causing execution to fail altogether when operation
nodes were present.