fix: remove unused code causing errors
The unused code, now removed, was causing a name error because the
block of which it was part was repurposed from pre-existing code and
part of the pre-existing code ended up not being removed as it should.

Some other lines that weren't causing errors but were not needed
either were also removed to avoid confusion and needless maintenance.