fix: implement missing type checks
Some missing type checks were causing bugs indirectly in different
modules in the source.

More specifically, when a number changed, if the value was the same,
even though the type was different, that change wasn't being treated
properly, because the app was only checking the value and reaching
the conclusion that nothing changed, because the value was the same
in the end. However, this is not the right thing to do. That is, we
must also take the type into account, because numeric values may have
the same value and still have different types.

For instance, 0, 0.0 and 0j all have the same value, although they
are an integer, a float and a complex number respectively. Because of
that, wherever a value was being checked to determine whether a change
took place, we also added a type check so this kind of change is also
taken into account.