misc: work on .gitattributes and normalize eol
The file now has text explaining:

- its purpose
- rules to be followed by people that may want to edit it

Many attributes were removed as well, since they were not strictly
needed. Files with .pyl extension are treated like python files
(the reason is also explained in the .gitattritues file).

The eol normalization refers to the introduction of end-of-line
normalization. This is performed by adding the changes in the
.gitattributes file with the "git add --renormalize" command,
just like explained on git's documention for the .gitattributes
file here: https://git-scm.com/docs/gitattributes

Such normalization causes existing tracked files with wrong CRLF/LF
line endings to be corrected, which also causes them to be added to
the commit.