fix: dialog panel growing offscreen
When there was too much text, the dialog was naturally growing beyond
the available space on the screen. The button to dismiss the dialog
was also ending up outside the screen, with no way for the user to
reach it, causing the user to be locked in the dialog, having to
restart the app in order to proceed.

In general, message dialogs are meant to show just a small amount
of text to inform or help the user make decisions. However, we were
allowing text from error reports to be displayed on it, instead of
the right practice of just using known predefined text. Text from
error reports can grow a lot when there are too many errors, which
was the cause of the problem.

Now, instead of reporting the errors on it, we just display an
informative small message and log the errors in the default logger
and our custom user logger. This way the user can check the error
report using the log files or the user log that can be accessed
on the app's GUI.