add: keyboard controls to confirm/dismiss dialogs
Added/set the ability to confirm/dismiss dialogs with the keyboard.

User can also navigate the available buttons using arrow keys and
the tab or shift+tab keys.

I use the terms "added/set" because part of the feature was already
implemented and just needed to be properly set in the existing
dialogs defined throughout the source.