add: automated system testing service
Added several new subpackages/modules so people can now perform
automated system testing sessions.

The main additions were:

- the nodezator/systemtesting subpackage
- the nodezator/systemtesting/reportviewer subpackage
- the nodezator/editing/playback/systemtesting.py module

The nodezator/systemtesting subpackage has subpackages/modules to
assist in system testing throughout the package.

One of such subpackages is nodezator/systemtesting/reportviewer,
which defines a system testing report viewer, a panel used to
visualize the results of a system testing session, that is, which tests
passed, failed or raised errors and much more.

This report viewer also has buttons to export the results in several
formats like .html, .pyl and .png.

The nodezator/editing/playback/systemtesting.py module contains a form
for users to set and trigger system testing sessions. In this form
people can pick which test cases to perform and the speed.

For now, the automated system testing feature/service has only 02
test cases, since we focused on finishing the system rather than
defining all the needed test cases and populating it with them.

In order to assist in this new feature, several other existing services
were updated/incremented and some smaller ones were created.

For users, the feature works like this: the user goes to the form in
the menubar option "Playback > Perform system testing session", sets
and triggers a system testing session, which starts performing the
test cases selected. All tests are executed automatically. During the
tests, the app moves automatically and the user must wait for the test
cases to finish executing. Optionally, the user can abort the testing
at any time by pressing the F7 key. If the user waits for the tests to
finish, a report of the results is shown.

The report can be revisited any time by selecting the menubar option
"Help > Show system testing report".

This commit is the result of several squashed commits.